Slightly improve tests, and fix a TokenMatcher bug that revealed. Move Token to utils.
